+++ packages/components/src/Animation/animations/Slide/index.ts
@@ -1,0 +1,2 @@
+export { default as SlideDown, SlideDownAnimation } from './SlideDown';
+export { default as SlideUp, SlideUpAnimation } from './SlideUp';