+++ packages/components/src/Animation/animations/index.ts
@@ -1,1 +1,2 @@
 export * from './Fade';
+export * from './Slide';